fix test class names
