+++ tests/Unit/Domain.LinnApps.Tests/StockLocatorServiceTests/WhenDeletingAndNotEmptyPallet.cs
@@ -1,6 +1,7 @@
 ﻿namespace Linn.Stores.Domain.LinnApps.Tests.StockLocatorServiceTests
 {
     using System;
+    using System.Collections.Generic;
     using System.Linq;
     using System.Linq.Expressions;
 
@@ -8,18 +9,34 @@
 
     using NUnit.Framework;
 
-    public class WhenDeletingAndNotEmptyPallet : ContextBase
+    public class WhenDeletingAndEmptyPallet : ContextBase
     {
         private readonly StockLocator toDelete = new StockLocator
                                                      {
                                                          Id = 1,
                                                          PalletNumber = 1
                                                      };
+
+        private readonly List<StockLocator> otherThingsInPallet = 
+            new List<StockLocator>
+                {
+                    new StockLocator
+                        {
+                            Id = 2,
+                            PalletNumber = 1
+                        },
+                    new StockLocator
+                        {
+                            Id = 3,
+                            PalletNumber = 1
+                        }
+                };
+
         [SetUp]
         public void SetUp()
         {
             this.StockLocatorRepository.FilterBy(Arg.Any<Expression<Func<StockLocator, bool>>>())
-                .Returns(Enumerable.Empty<StockLocator>().AsQueryable());
+                .Returns(this.otherThingsInPallet.AsQueryable());
             this.Sut.DeleteStockLocator(this.toDelete);
         }
 
@@ -30,9 +47,9 @@
         }
 
         [Test]
-        public void ShouldNotUpdatePallets()
+        public void ShouldUpdatePallets()
         {
-            this.StoresPalletRepository.DidNotReceive()
+            this.StoresPalletRepository.Received()
                 .FilterBy(Arg.Any<Expression<Func<StoresPallet, bool>>>());
         }
     }
